pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mirastack-sdk"
version = "0.1.0"
description = "Python SDK for building MIRASTACK plugins"
readme = "README.md"
license = "AGPL-3.0-or-later"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.72.0",
    "grpcio-tools>=1.72.0",
    "protobuf>=5.29.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
    "grpc-stubs>=1.53",
]
